admission/webhook: refactor to webhook = generic-webhook + source + dispatcher
