+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/validating/plugin_test.go
@@ -36,6 +36,7 @@ import (
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apiserver/pkg/admission"
+	"k8s.io/apiserver/pkg/admission/configuration"
 	"k8s.io/apiserver/pkg/admission/plugin/webhook/config"
 	"k8s.io/apiserver/pkg/admission/plugin/webhook/testcerts"
 	"k8s.io/apiserver/pkg/authentication/user"
@@ -47,7 +48,7 @@ type fakeHookSource struct {
 	err   error
 }
 
-func (f *fakeHookSource) Webhooks() *registrationv1beta1.ValidatingWebhookConfiguration {
+func (f *fakeHookSource) Webhooks() []registrationv1beta1.Webhook {
 	if f.err != nil {
 		return nil
 	}
@@ -56,7 +57,11 @@ func (f *fakeHookSource) Webhooks() *registrationv1beta1.ValidatingWebhookConfig
 			f.hooks[i].NamespaceSelector = &metav1.LabelSelector{}
 		}
 	}
-	return &registrationv1beta1.ValidatingWebhookConfiguration{Webhooks: f.hooks}
+	return f.hooks
+}
+
+func (f *fakeHookSource) HasSynched() bool {
+	return true
 }
 
 func (f *fakeHookSource) Run(stopCh <-chan struct{}) {}
@@ -115,7 +120,7 @@ func (c urlConfigGenerator) ccfgURL(urlPath string) registrationv1beta1.WebhookC
 	}
 }
 
-// TestValidate tests that ValidatingAdmissionWebhook#Validate works as expected
+// TestValidate tests that ValidatingWebhook#Validate works as expected
 func TestValidate(t *testing.T) {
 	scheme := runtime.NewScheme()
 	v1beta1.AddToScheme(scheme)
@@ -128,7 +133,7 @@ func TestValidate(t *testing.T) {
 	if err != nil {
 		t.Fatalf("this should never happen? %v", err)
 	}
-	wh, err := NewValidatingAdmissionWebhook(nil)
+	wh, err := NewValidatingAdmissionWebhook(nil, configuration.NewValidatingWebhookConfigurationManager)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -391,7 +396,7 @@ func TestValidate(t *testing.T) {
 			continue
 		}
 		t.Run(name, func(t *testing.T) {
-			wh.hookSource = &tt.hookSource
+			wh.Webhook.hookSource = &tt.hookSource
 			err = wh.Validate(admission.NewAttributesRecord(&object, &oldObject, kind, namespace, name, resource, subResource, operation, &userInfo))
 			if tt.expectAllow != (err == nil) {
 				t.Errorf("expected allowed=%v, but got err=%v", tt.expectAllow, err)
@@ -409,7 +414,7 @@ func TestValidate(t *testing.T) {
 	}
 }
 
-// TestValidateCachedClient tests that ValidatingAdmissionWebhook#Validate should cache restClient
+// TestValidateCachedClient tests that ValidatingWebhook#Validate should cache restClient
 func TestValidateCachedClient(t *testing.T) {
 	scheme := runtime.NewScheme()
 	v1beta1.AddToScheme(scheme)
@@ -422,7 +427,7 @@ func TestValidateCachedClient(t *testing.T) {
 	if err != nil {
 		t.Fatalf("this should never happen? %v", err)
 	}
-	wh, err := NewValidatingAdmissionWebhook(nil)
+	wh, err := NewValidatingAdmissionWebhook(nil, configuration.NewValidatingWebhookConfigurationManager)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -431,7 +436,7 @@ func TestValidateCachedClient(t *testing.T) {
 		t.Fatalf("cannot create client manager: %v", err)
 	}
 	cm.SetServiceResolver(fakeServiceResolver{base: *serverURL})
-	wh.clientManager = cm
+	wh.Webhook.clientManager = cm
 	wh.SetScheme(scheme)
 	namespace := "webhook-test"
 	wh.namespaceMatcher.NamespaceLister = fakeNamespaceLister{map[string]*corev1.Namespace{
@@ -551,11 +556,11 @@ func TestValidateCachedClient(t *testing.T) {
 
 	for _, testcase := range cases {
 		t.Run(testcase.name, func(t *testing.T) {
-			wh.hookSource = &testcase.hookSource
+			wh.Webhook.hookSource = &testcase.hookSource
 			authInfoResolverCount := new(int32)
 			r := newFakeAuthenticationInfoResolver(authInfoResolverCount)
-			wh.clientManager.SetAuthenticationInfoResolver(r)
-			if err = wh.clientManager.Validate(); err != nil {
+			wh.Webhook.clientManager.SetAuthenticationInfoResolver(r)
+			if err = wh.Webhook.clientManager.Validate(); err != nil {
 				t.Fatal(err)
 			}
